Move iterator utilities into their own module
