+++ pkg/bloomgateway/processor.go
@@ -11,6 +11,7 @@ import (
 	"github.com/grafana/dskit/concurrency"
 	"github.com/grafana/dskit/multierror"
 
+	iter "github.com/grafana/loki/v3/pkg/iter/v2"
 	v1 "github.com/grafana/loki/v3/pkg/storage/bloom/v1"
 	"github.com/grafana/loki/v3/pkg/storage/config"
 	"github.com/grafana/loki/v3/pkg/storage/stores/shipper/bloomshipper"
@@ -149,7 +150,7 @@ func (p *processor) processBlock(_ context.Context, bq *bloomshipper.CloseableBl
 	}
 
 	tokenizer := v1.NewNGramTokenizer(schema.NGramLen(), schema.NGramSkip())
-	iters := make([]v1.PeekingIterator[v1.Request], 0, len(tasks))
+	iters := make([]iter.PeekingIterator[v1.Request], 0, len(tasks))
 
 	for _, task := range tasks {
 		// NB(owen-d): can be helpful for debugging, but is noisy
@@ -162,7 +163,7 @@ func (p *processor) processBlock(_ context.Context, bq *bloomshipper.CloseableBl
 		// 	sp.LogKV("process block", blockID, "series", len(task.series))
 		// }
 
-		it := v1.NewPeekingIter(task.RequestIter(tokenizer))
+		it := iter.NewPeekingIter(task.RequestIter(tokenizer))
 		iters = append(iters, it)
 	}
